refactor lp12-parentR2 so that lp17 can use it
