+++ tests/unit/libpluto/lp12-parentR2/parentR2_head.c
@@ -1,0 +1,41 @@
+#define LEAK_DETECTIVE
+#define AGGRESSIVE 1
+#define XAUTH
+#define MODECFG
+#define DEBUG 1
+#define PRINT_SA_DEBUG 1
+#define USE_KEYRR 1
+
+#include <pcap.h>
+
+#include "constants.h"
+#include "oswalloc.h"
+#include "whack.h"
+#include "../../programs/pluto/rcv_whack.h"
+
+#include "../../programs/pluto/connections.c"
+
+#include "whackmsgtestlib.c"
+#include "seam_timer.c"
+#include "seam_fakevendor.c"
+#include "seam_pending.c"
+#include "seam_ikev1.c"
+#include "seam_crypt.c"
+#include "seam_kernel.c"
+#include "seam_rnd.c"
+#include "seam_log.c"
+#include "seam_xauth.c"
+#include "seam_host_jamesjohnson.c"
+#include "seam_terminate.c"
+#include "seam_x509.c"
+#include "seam_spdbstruct.c"
+#include "seam_demux.c"
+#include "seam_commhandle.c"
+#include "seam_whack.c"
+#include "seam_exitlog.c"
+#include "seam_natt.c"
+#include "seam_dnskey.c"
+#include "seam_kernelalgs.c"
+
+#include "seam_gr_sha1_group14.c"
+
